+++ src/api/appendToSheetData.ts
@@ -6,11 +6,7 @@ const SPREADSHEET_ID = process.env.NEXT_PUBLIC_SPREADSHEET_ID || "";
 const doc = new GoogleSpreadsheet(SPREADSHEET_ID, jwt);
 
 export const appendToSheet = async (data: Record<string, string>) => {
-    try {
-        await doc.loadInfo(); // loads document properties and worksheets
-        const sheet = doc.sheetsByIndex[0]; // assuming the first sheet is where you want to append data
-        await sheet.addRow(data);
-    } catch (error) {
-        return error;
-    }
+    await doc.loadInfo(); // loads document properties and worksheets
+    const sheet = doc.sheetsByIndex[0]; // assuming the first sheet is where you want to append data
+    await sheet.addRow(data);
 };
