+++ src/components/shared/WriteUs/NotificationModal.tsx
@@ -1,0 +1,1 @@
+export const NotififcationModal = () => {};
